refactor: move evaluation evidence into package
