+++ src/evolve/evaluation/execution.py
@@ -12,7 +12,7 @@ from ..config import evaluator_boolean, evaluator_sampling, experiment_id, load_
 from ..git import evaluator_tree, git, git_stdout
 from ..host_runtime import clean_python_env
 from ..runtime import OwnedResult, attempt_dir, next_attempt, owned_attempt_id, run_owned
-from ..task_vectors import trial_results, validate_task_vector
+from .evidence import trial_results, validate_task_vector
 from .identity import effective_task_set_identity
 from .results import EvaluationRecord, Outcome, classify_evaluation
